Move CacheKey to its own source file
Necessary for JDK10 compiles as CacheKey is referred to from other
files.  JDK10 compiles require it to be in its own file.
